* Fixed bug that Alister discovered.
When loading an image we assumed an origin in the VtkStructured points that resulted in the image centre for a slice z being at the (0,0,z). The image provided by Alister broke this assumption.

So to fix it we force this assumption and hope that it doesn't break anything else :)